requirements_server.txt
Flask
argon2-cffi
psycopg
python-dotenv
